M4 L5 UPDATE

+++ module4/lesson05_orm/exercise_product_manager/src/main/java/com/product/repository/ProductRepository.java
@@ -9,7 +9,7 @@ public interface ProductRepository {
 
     void create(Product product);
 
-    Product findById(int id);
+    Product findById(long id);
 
-    void delete(int id);
+    void delete(long id);
 }
